fix: address PRRT_kwDOSJAM6s6S7WnD — restore request schema ownership
